requirements.txt
django==3.2.6
djangorestframework~=3.12.4
pip~=21.2.4
pytz~=2021.1
sqlparse~=0.4.1
asgiref~=3.4.1
setuptools~=57.4.0
